requirements.txt
pandas==2.2.2
pyyaml==6.0.2
tqdm==4.66.5
requests==2.32.3
beautifulsoup4==4.12.3
python-dateutil==2.9.0
snscrape==0.7.0.20230622
langdetect==1.0.9
fasttext-langdetect
unidecode==1.3.8
chardet==5.2.0
kagglehub[pandas-datasets]